feat(backend): changing delta from double to int

+++ backend/src/main/java/de/kaliburg/morefair/game/ladder/model/dto/LadderTickDto.java
@@ -12,5 +12,5 @@ public class LadderTickDto {
 
   @NonNull
   @Builder.Default
-  Double delta = 1d;
+  Integer delta = 1;
 }
